requirements.txt
fastapi>=0.115,<1
pymongo[srv]>=4.11,<5
python-dotenv>=1.0,<2
httpx>=0.28,<1
yfinance>=0.2.54,<1
